gestion_negocio/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asn1crypto==1.5.1
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
black==25.1.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
limits==4.0.1
Mako==1.3.9
MarkupSafe==3.0.2
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
rsa==4.9
scramp==1.4.5
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.34.0
wrapt==1.17.2
